Inlining is-object dependency
Given the popularity of this pacakge, I thought it migth be helpful to
remove a 1-liner depedencency. Saves a tiny bit of time everywhere, and
perhaps promotes the idea that throwaway npm dependencies aren't all
that necessary.

I understand and will not be offended if this goes against philosophy.
We're cool.

If it is accepted I will propose the same change on
`is-urlsearchparams`.